+++ src/main/java/de/samply/db/repository/ProjectBridgeheadRepository.java
@@ -6,10 +6,14 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
+import java.util.Set;
 
 @Repository
 public interface ProjectBridgeheadRepository extends JpaRepository<ProjectBridgehead, Long> {
 
     Optional<ProjectBridgehead> findFirstByBridgeheadAndProject(String bridgehead, Project project);
+
+    Set<ProjectBridgehead> findByProject(Project project);
+
 
 }