+++ src/main/java/de/samply/notification/smtp/EmailContext.java
@@ -1,0 +1,16 @@
+package de.samply.notification.smtp;
+
+import lombok.Getter;
+import org.springframework.stereotype.Component;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@Component
+public class EmailContext {
+
+    @Getter
+    private Map<String, String> keyValues = new HashMap<>();
+
+
+}