+++ src/main/java/de/samply/notification/smtp/EmailServiceException.java
@@ -1,0 +1,9 @@
+package de.samply.notification.smtp;
+
+public class EmailServiceException extends Exception {
+
+    public EmailServiceException(String message) {
+        super(message);
+    }
+
+}