+++ src/main/java/de/samply/notification/smtp/MessageSubject.java
@@ -1,0 +1,4 @@
+package de.samply.notification.smtp;
+
+public record MessageSubject (String message, String subject) {
+}