+++ src/main/java/de/samply/notification/smtp/TemplateSubject.java
@@ -1,0 +1,4 @@
+package de.samply.notification.smtp;
+
+public record TemplateSubject (String template, String subject) {
+}